修复了SqlServer2005的分页bug 优化了JdbcExecutor接口

+++ jpacker/src/test/java/jpacker/sqlserver.sql
@@ -1,5 +1,5 @@
 create table users(
-	id			bigint 			not null primary key	IDENTITY(1,1),
+	id			int 			not null primary key	IDENTITY(1,1),
 	username	varchar(20)		not null,
 	password	varchar(20)		not null,
 	realname	varchar(20)		null,
